chore: migrate the UriParsingHelper to Prism.Core
