--- src/Wpf/Prism.Wpf/Common/UriParsingHelper.cs
@@ -1,55 +1,0 @@
-using System;
-using Prism.Navigation;
-using Prism.Regions;
-
-namespace Prism.Common
-{
-    /// <summary>
-    /// Helper class for parsing <see cref="Uri"/> instances.
-    /// </summary>
-    public static class UriParsingHelper
-    {
-        /// <summary>
-        /// Gets the query part of <paramref name="uri"/>.
-        /// </summary>
-        /// <param name="uri">The Uri.</param>
-        public static string GetQuery(Uri uri)
-        {
-            return EnsureAbsolute(uri).Query;
-        }
-
-        /// <summary>
-        /// Gets the AbsolutePath part of <paramref name="uri"/>.
-        /// </summary>
-        /// <param name="uri">The Uri.</param>
-        public static string GetAbsolutePath(Uri uri)
-        {
-            return EnsureAbsolute(uri).AbsolutePath;
-        }
-
-        /// <summary>
-        /// Parses the query of <paramref name="uri"/> into a dictionary.
-        /// </summary>
-        /// <param name="uri">The URI.</param>
-        public static INavigationParameters ParseQuery(Uri uri)
-        {
-            var query = GetQuery(uri);
-
-            return new NavigationParameters(query);
-        }
-
-        private static Uri EnsureAbsolute(Uri uri)
-        {
-            if (uri.IsAbsoluteUri)
-            {
-                return uri;
-            }
-
-            if ((uri != null) && !uri.OriginalString.StartsWith("/", StringComparison.Ordinal))
-            {
-                return new Uri("http://localhost/" + uri, UriKind.Absolute);
-            }
-            return new Uri("http://localhost" + uri, UriKind.Absolute);
-        }
-    }
-}
